requirements.txt
google-generativeai
streamlit
python-dotenv
Pillow # for PIL(python image library)
